requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.2
boto==2.48.0
boto3==1.5.7
botocore==1.8.21
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
docutils==0.14
enum34==1.1.6
gensim==3.2.0
html5lib==1.0.1
idna==2.6
jmespath==0.9.3
Markdown==2.6.10
mecab-python3==0.7
mojimoji==0.0.6
numpy==1.13.3
protobuf==3.5.1
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
smart-open==1.5.6
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.13
